+++ src/collective/cover/testing.py
@@ -99,6 +99,7 @@ class Fixture(PloneSandboxLayer):
         portal['my-image2'].setImage(generate_jpeg(50, 50))
         portal['my-file'].setFile(loadFile('lorem_ipsum.txt'))
         portal['my-file'].reindexObject()
+        portal['my-news-item'].setImage(generate_jpeg(50, 50))
         portal_workflow = portal.portal_workflow
         portal_workflow.setChainForPortalTypes(['Collection'],
                                                ['plone_workflow'],)