+++ docs/CHANGES.rst
@@ -4,6 +4,10 @@ There's a frood who really knows where his towel is
 1.0a5 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Added any content type support for banner tile. If it has an `image`
+  or `getImage` attribute, displays that image. And always adds
+  a link (closes `#241`_). [marcosfromero]
+
 - Move CSS to registry_css (closes `#244`_). [agnogueira]
 
 - Collection tile now fulfills the configured image size (fixes `#239`_).
@@ -218,3 +222,4 @@ There's a frood who really knows where his towel is
 .. _`#206`: https://github.com/collective/collective.cover/issues/206
 .. _`#218`: https://github.com/collective/collective.cover/issues/218
 .. _`#239`: https://github.com/collective/collective.cover/issues/239
+.. _`#241`: https://github.com/collective/collective.cover/issues/241